pyproject.toml
[project]
name = "gatus-operator"
version = "0.2.0"
description = "Kubernetes operator for Gatus"
requires-python = ">=3.8"
dependencies = [
    "kubernetes==33.1.0",
    "ruamel.yaml==0.18.6",
] 

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0",
]
